Move base classes into __init__ and flake8 code cleanup
