+++ tests/test_device.py
@@ -1,6 +1,6 @@
 import binascii
 from insteonplm.address import Address
-from insteonplm.devices import Device
+import insteonplm.devices
 from insteonplm.devices.generalController import GeneralController
 from insteonplm.devices.dimmableLightingControl import DimmableLightingControl
 from insteonplm.devices.switchedLightingControl import SwitchedLightingControl
@@ -8,7 +8,7 @@ from .mockPLM import MockPLM
 
 def test_create_device():
     plm = MockPLM()
-    device = Device.create(plm, '112233', 0x01, 0x0d, None)
+    device = insteonplm.devices.create(plm, '112233', 0x01, 0x0d, None)
     assert device.id == '112233'
     assert isinstance(device, DimmableLightingControl)
 
@@ -17,6 +17,6 @@ def test_create_device_from_bytearray():
     target = bytearray()
     target.append(0x01)
     target.append(0x0d)
-    device = Device.create(plm, '112233', target[0], target[1], None)
+    device = insteonplm.devices.create(plm, '112233', target[0], target[1], None)
     assert device.id == '112233'
     assert isinstance(device, DimmableLightingControl)
